Flushes all IPv6 routes after configuring an interface with an IPv4 address
ip link -4 link set ifname up has proven to add unwanted IPv6 routes found on the host to the container.
We want these routes to be added when configuring the IPv6 interface, not the IPv4. This was causing duplicate
IPv6 routes on two differnt interfaces, which was causing problems on our internal tools. This flushes all IPv6 routes
after configuring the IPv4 interface to prevent duplicate routes.